Fixes for 64 bit compilation. Contributed by alexr
git-svn-id: https://csp.svn.sourceforge.net/svnroot/csp/trunk@2222 7152ba15-5bd8-0310-b828-e896570a432b

+++ csp/tools/layout/LayoutNodes.h
@@ -89,7 +89,7 @@ public:
 	virtual void accept(LayoutNodeVisitor &v)=0;
 	virtual void traverse(LayoutNodeVisitor &v)=0;
 
-	unsigned int id() { return reinterpret_cast<unsigned int>(this); }
+	size_t id() { return reinterpret_cast<size_t>(this); }
 
 	virtual void transform(osg::Matrix const &t) {
 		osg::Matrix m = _move->getMatrix();
